% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolicQspray.R
\docType{methods}
\name{as.symbolicQspray}
\alias{as.symbolicQspray}
\alias{as.symbolicQspray,character-method}
\alias{as.symbolicQspray,qspray-method}
\alias{as.symbolicQspray,ratioOfQsprays-method}
\alias{as.symbolicQspray,symbolicQspray-method}
\alias{as.symbolicQspray,numeric-method}
\alias{as.symbolicQspray,bigz-method}
\alias{as.symbolicQspray,bigq-method}
\title{Coercion to a 'symbolicQspray' object}
\usage{
\S4method{as.symbolicQspray}{character}(x)

\S4method{as.symbolicQspray}{qspray}(x)

\S4method{as.symbolicQspray}{ratioOfQsprays}(x)

\S4method{as.symbolicQspray}{symbolicQspray}(x)

\S4method{as.symbolicQspray}{numeric}(x)

\S4method{as.symbolicQspray}{bigz}(x)

\S4method{as.symbolicQspray}{bigq}(x)
}
\arguments{
\item{x}{a \code{symbolicQspray} object or an object for which
\code{\link[ratioOfQsprays]{as.ratioOfQsprays}} is applicable}
}
\value{
A \code{symbolicQspray} object.
}
\description{
Coercion to a 'symbolicQspray' object
}
\examples{
as.symbolicQspray(2)
as.symbolicQspray("1/3")
}
