% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetricPolynomials.R
\docType{methods}
\name{compactSymmetricQspray}
\alias{compactSymmetricQspray}
\alias{compactSymmetricQspray,symbolicQspray,logical-method}
\alias{compactSymmetricQspray,symbolicQspray,missing-method}
\title{Compact symmetric qspray}
\usage{
\S4method{compactSymmetricQspray}{symbolicQspray,logical}(qspray, check)

\S4method{compactSymmetricQspray}{symbolicQspray,missing}(qspray, check)
}
\arguments{
\item{qspray}{a \code{symbolicQspray} object which should correspond to a
symmetric polynomial}

\item{check}{Boolean, whether to check the symmetry}
}
\value{
A character string.
}
\description{
Prints a symmetric \code{symbolicQspray} polynomial as a linear
  combination of the monomial symmetric polynomials.
}
\seealso{
\code{\link[qspray]{MSPcombination}}
}
