% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{derivSymbolicQspray}
\alias{derivSymbolicQspray}
\title{Partial derivative}
\usage{
derivSymbolicQspray(Qspray, i, derivative = 1)
}
\arguments{
\item{Qspray}{object of class \code{symbolicQspray}}

\item{i}{integer, the dimension to differentiate with respect to, e.g.
\code{2} to differentiate w.r.t. \eqn{y}}

\item{derivative}{positive integer, how many times to differentiate}
}
\value{
A \code{symbolicQspray} object.
}
\description{
Partial derivative of a \code{symbolicQspray} polynomial.
}
