% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{getConstantTerm}
\alias{getConstantTerm}
\alias{getConstantTerm,symbolicQspray-method}
\title{Get the constant term of a 'symbolicQspray' polynomial}
\usage{
\S4method{getConstantTerm}{symbolicQspray}(qspray)
}
\arguments{
\item{qspray}{a \code{symbolicQspray} object}
}
\value{
A \code{ratioOfQsprays} object.
}
\description{
Get the constant term of a \code{symbolicQspray} polynomial.
}
