% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isConstant}
\alias{isConstant}
\alias{isConstant,symbolicQspray-method}
\title{Whether a 'symbolicQspray' polynomial is constant}
\usage{
\S4method{isConstant}{symbolicQspray}(x)
}
\arguments{
\item{x}{a \code{symbolicQspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{symbolicQspray} object defines a constant
  polynomial.
}
