% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isQone}
\alias{isQone}
\alias{isQone,symbolicQspray-method}
\title{Whether a 'symbolicQspray' polynomial is the unit polynomial}
\usage{
\S4method{isQone}{symbolicQspray}(qspray)
}
\arguments{
\item{qspray}{a \code{symbolicQspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{symbolicQspray} object defines the unit
  polynomial.
}
