% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isQzero}
\alias{isQzero}
\alias{isQzero,symbolicQspray-method}
\title{Whether a 'symbolicQspray' polynomial is null}
\usage{
\S4method{isQzero}{symbolicQspray}(qspray)
}
\arguments{
\item{qspray}{a \code{symbolicQspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{symbolicQspray} object defines the zero
  polynomial.
}
