% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isUnivariate}
\alias{isUnivariate}
\alias{isUnivariate,symbolicQspray-method}
\title{Whether a 'symbolicQspray' polynomial is univariate}
\usage{
\S4method{isUnivariate}{symbolicQspray}(x)
}
\arguments{
\item{x}{a \code{symbolicQspray} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{symbolicQspray} object defines a
  univariate polynomial.
}
\note{
It is considered that a constant \code{symbolicQspray} is univariate.
}
