% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{numberOfParameters}
\alias{numberOfParameters}
\title{Number of parameters}
\usage{
numberOfParameters(Qspray)
}
\arguments{
\item{Qspray}{a \code{symbolicQspray} object}
}
\value{
An integer, the number of parameters involved in (the
  coefficients of) \code{Qspray}.
}
\description{
Number of parameters of a \code{symbolicQspray} polynomial,
  i.e. the number of variables occurring in its coefficients.
}
\examples{
JP <- JacobiPolynomial(4) # Jacobi polynomials have two parameters
numberOfParameters(JP)
}
