% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{numberOfTerms}
\alias{numberOfTerms}
\alias{numberOfTerms,symbolicQspray-method}
\title{Number of terms in a 'symbolicQspray' polynomial}
\usage{
\S4method{numberOfTerms}{symbolicQspray}(qspray)
}
\arguments{
\item{qspray}{a \code{symbolicQspray} object}
}
\value{
An integer.
}
\description{
Number of terms in the polynomial defined by a
  \code{symbolicQspray} object.
}
