% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{numberOfVariables}
\alias{numberOfVariables}
\alias{numberOfVariables,symbolicQspray-method}
\title{Number of variables of a 'symbolicQspray' polynomial}
\usage{
\S4method{numberOfVariables}{symbolicQspray}(x)
}
\arguments{
\item{x}{a \code{symbolicQspray} object}
}
\value{
An integer.
}
\description{
Number of variables involved in a \code{symbolicQspray} object.
}
\note{
The number of variables in the \code{symbolicQspray} object
  \code{Qlone(d)} is \code{d}, not \code{1}.
}
\seealso{
\code{\link{involvedVariables}}.
}
