% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{Qlone}
\alias{Qlone}
\title{Polynomial variable}
\usage{
Qlone(n)
}
\arguments{
\item{n}{nonnegative integer, the index of the variable}
}
\value{
A \code{qspray} object.
}
\description{
Create a polynomial variable for a \code{symbolicQspray}.
}
\examples{
Qlone(2)
}
