% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{Qone}
\alias{Qone}
\title{The unit 'symbolicQspray' polynomial}
\usage{
Qone()
}
\value{
A \code{symbolicQspray} object.
}
\description{
Returns the \code{symbolicQspray} polynomial identically
  equal to 1.
}
