% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{Qzero}
\alias{Qzero}
\title{The null 'symbolicQspray' polynomial}
\usage{
Qzero()
}
\value{
A \code{symbolicQspray} object.
}
\description{
Returns the \code{symbolicQspray} polynomial
  identically equal to 0.
}
