% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{rSymbolicQspray}
\alias{rSymbolicQspray}
\title{Random 'symbolicQspray'}
\usage{
rSymbolicQspray()
}
\value{
A \code{symbolicQspray} object.
}
\description{
Generates a random \code{symbolicQspray} object.
}
