% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolicQspray.R
\docType{methods}
\name{showCoefficient}
\alias{showCoefficient}
\alias{showCoefficient,symbolicQspray-method}
\title{Function which prints the coefficients}
\usage{
\S4method{showCoefficient}{symbolicQspray}(x)
}
\arguments{
\item{x}{a \code{symbolicQspray} object}
}
\value{
A function which associates a character string to the coefficients.
}
\description{
This method is for internal usage. For a \code{qspray}
  object, it returns a function which takes as argument a coefficient
  and which returns a string displaying this coefficient surrounded by
  curly braces.
}
