% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showSymbolicQspray}
\alias{showSymbolicQspray}
\title{Print a 'symbolicQspray' object}
\usage{
showSymbolicQspray(showRatioOfQsprays, showMonomial)
}
\arguments{
\item{showRatioOfQsprays}{a function which prints a \code{ratioOfQsprays}
object}

\item{X}{a string, usually a letter, to denote the non-indexed variables}
}
\value{
A function which prints a \code{symbolicQspray} object.
}
\description{
Print a \code{symbolicQspray} object given a function to print
  a \code{ratioOfQsprays} object.
}
\seealso{
\code{\link{showSymbolicQsprayX1X2X3}},
  \code{\link{showSymbolicQsprayXYZ}}.
}
