% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showSymbolicQsprayCanonical}
\alias{showSymbolicQsprayCanonical}
\title{Print a 'symbolicQspray' object}
\usage{
showSymbolicQsprayCanonical(a = "a", X = "X", quotientBar = "  \%//\%  ", ...)
}
\arguments{
\item{a}{a string, usually a letter, to denote the unindexed variables
of the \code{ratioOfQsprays} coefficients}

\item{X}{a string, usually a letter, to denote the unindexed variables}

\item{quotientBar}{a string for the quotient bar between the numerator and
the denominator of a \code{ratioOfQsprays} object, including surrounding
spaces, e.g. \code{"/"}}

\item{...}{arguments other than \code{var} passed to
\code{showSymbolicQspray} (currently there is no such argument)}
}
\value{
A function which prints \code{symbolicQspray} objects.
}
\description{
Print a \code{symbolicQspray} object.
}
\note{
The \code{show} method for \code{symbolicQspray} objects uses
  \code{showSymbolicQsprayCanonical("a", "X", "  \%//\%  ")}.
}
\examples{
Qspray <- rSymbolicQspray()
showSymbolicQsprayCanonical(quotientBar = " / ")
}
