% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showSymbolicQsprayOption<-}
\alias{showSymbolicQsprayOption<-}
\title{Set show option to a 'qspray' object}
\usage{
showSymbolicQsprayOption(x, which) <- value
}
\arguments{
\item{x}{a \code{qspray} object}

\item{which}{which option to set; this can be \code{"x"},
\code{"showMonomial"}, or \code{"showQspray"}}

\item{value}{the value of the option}
}
\value{
This returns the updated \code{qspray}.
}
\description{
Set show option to a \code{qspray} object
}
\examples{
Qspray <- rSymbolicQspray()
showSymbolicQsprayOption(Qspray, "a") <- "x"
showSymbolicQsprayOption(Qspray, "X") <- "A"
showSymbolicQsprayOption(Qspray, "quotientBar") <- " / "
Qspray
}
