% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolicQspray.R
\name{symbolicQspray-unary}
\alias{symbolicQspray-unary}
\alias{+,symbolicQspray,missing-method}
\alias{-,symbolicQspray,missing-method}
\title{Unary operators for 'symbolicQspray objects}
\usage{
\S4method{+}{symbolicQspray,missing}(e1, e2)

\S4method{-}{symbolicQspray,missing}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{symbolicQspray}}

\item{e2}{nothing}
}
\value{
A \code{symbolicQspray} object.
}
\description{
Unary operators for \code{symbolicQspray} objects.
}
