% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{symbolicQspray_from_list}
\alias{symbolicQspray_from_list}
\title{(internal) Make a 'symbolicQspray' object from a list}
\usage{
symbolicQspray_from_list(x)
}
\arguments{
\item{x}{list returned by the Rcpp function
\code{returnSymbolicQspray}}
}
\value{
A \code{symbolicQspray} object.
}
\description{
This function is for internal usage. It is exported because
  it is also used for internal usage in others packages.
}
