% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{withAttributes}
\alias{withAttributes}
\title{Set show options to a 'symbolicQspray' object}
\usage{
withAttributes(Qspray, a = "a", X = "X", quotientBar = " \%//\% ")
}
\arguments{
\item{Qspray}{a \code{symbolicQspray} object}

\item{a}{value for the \code{"a"} attribute}

\item{X}{value for the \code{"X"} attribute}

\item{quotientBar}{value for the \code{"quotientBar"} attribute}
}
\value{
The input \code{symbolicQspray} object with new attributes.
}
\description{
Set some attributes to a \code{symbolicQspray} object
  to control the way it is displayed. See#' the note in the
  documentation of \code{\link[showSymbolicQsprayCanonical]} for details.
}
\examples{
( Qspray <- rSymbolicQspray() )
withAttributes(Qspray, a = "x", X = "A", quotientBar = " / ")
}
