% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evalSymbolicQspray}
\alias{evalSymbolicQspray}
\title{Evaluation of a 'symbolicQspray' object}
\usage{
evalSymbolicQspray(Qspray, a = NULL, X = NULL)
}
\arguments{
\item{Qspray}{a \code{symbolicQspray} object}

\item{a}{vector of values to be substituted to the exterior variables:
these values must be coercable to \code{bigq} numbers}

\item{X}{vector of values to be substituted to the main variables: these
values must be coercable to \code{bigq} numbers}
}
\value{
If both \code{a} and \code{X} are \code{NULL}, this returns the
  input \code{symbolicQspray} object; otherwise, if \code{a} is not
  \code{NULL}, this returns a \code{qspray} object, and if
  \code{X} is not \code{NULL}, this returns a \code{ratioOfQsprays} object.
}
\description{
Evaluates a \code{symbolicQspray} object by substituting some
  values to the exterior variables or the main variables or both.
}
\examples{
library(symbolicQspray)
a1 <- qlone(1)
a2 <- qlone(2)
X1 <- Qlone(1)
X2 <- Qlone(2)
X3 <- Qlone(3)
( Qspray <- (a1 + 2)*X1^2*X2 + (a2/(a1^2+a2))*X1*X2*X3 )
a <- c(2, 3)
X <- c(4, 3, 2)
( qspray <- evalSymbolicQspray(Qspray, a = a) )
( roq <- evalSymbolicQspray(Qspray, X = X) )
evalSymbolicQspray(Qspray, a = a, X = X)
evalQspray(qspray, X)
evalRatioOfQsprays(roq, a)
}
