% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evalSymbolicQspray}
\alias{evalSymbolicQspray}
\title{Evaluation of a 'symbolicQspray' object}
\usage{
evalSymbolicQspray(Qspray, a = NULL, X = NULL)
}
\arguments{
\item{Qspray}{a \code{symbolicQspray} object}

\item{a}{values to be substituted to the inner variables}

\item{X}{values to be substituted to the main variables}
}
\value{
If both \code{a} and \code{X} are \code{NULL}, this returns the
  input \code{symbolicQspray} object; otherwise, if \code{a} is not
  \code{NULL}, this returns a \code{qspray} object, and if \code{X} is not
  \code{NULL}, this returns a \code{ratioOfQsprays} object.
}
\description{
Evaluates a \code{symbolicQspray} object by substituting some
  values to the inner variables or the main variables or both.
}
\examples{
library(symbolicQspray)
a1 <- qlone(1)
a2 <- qlone(2)
X1 <- Qlone(1)
X2 <- Qlone(2)
X3 <- Qlone(3)
( Qspray <- (a1 + 2)*X1^2*X2 + (a2/(a1^2+a2))*X1*X2*X3 )
a <- c(2, 3)
X <- c(4, 3, 2)
( qspray <- evalSymbolicQspray(Qspray, a = a) )
( roq <- evalSymbolicQspray(Qspray, X = X) )
evalSymbolicQspray(Qspray, a = a, X = X)
evalQspray(qspray, X)
evalRatioOfQsprays(roq, a)
# More generally, X is a list of `ratioOfQsprays` objects
X <- list(qlone(1)/(1+qlone(1)), qlone(2)/(qlone(1)^2), qlone(3)/qlone(1))
evalSymbolicQspray(Qspray, a = a, X = X)   # ratioOfQsprays
( Q <- evalSymbolicQspray(Qspray, a = a) ) # symbolicQspray
evalSymbolicQspray(Q, X = X)               # ratioOfQsprays
}
