% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{Qlone}
\alias{Qlone}
\title{Polynomial variable}
\usage{
Qlone(n)
}
\arguments{
\item{n}{positive integer, the index of the variable}
}
\value{
A \code{symbolicQspray} object.
}
\description{
Creates a polynomial variable for a \code{symbolicQspray}.
}
\examples{
X <- Qlone(1)
Y <- Qlone(2)
(X + Y)^2
Qlone(0) == 1
}
