% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showSymbolicQsprayCanonical}
\alias{showSymbolicQsprayCanonical}
\title{Print a 'symbolicQspray' object}
\usage{
showSymbolicQsprayCanonical(a = "a", X = "X", quotientBar = " \%//\% ", ...)
}
\arguments{
\item{a}{a string, usually a letter, to denote the unindexed variables
of the \code{ratioOfQsprays} coefficients}

\item{X}{a string, usually a letter, to denote the unindexed variables}

\item{quotientBar}{a string for the quotient bar between the numerator and
the denominator of a \code{ratioOfQsprays} object, including surrounding
spaces, e.g. \code{"/"}}

\item{...}{arguments other than \code{var} passed to
\code{showSymbolicQspray} (currently there is no such argument)}
}
\value{
A function which prints \code{symbolicQspray} objects.
}
\description{
Print a \code{symbolicQspray} object.
}
\note{
The \code{show} method for \code{symbolicQspray} objects uses
  \code{showSymbolicQsprayCanonical("a", "X", " \%//\% ")} by default.
  But this can be controlled as follows. If a
  \code{symbolicQspray} object has an attribute \code{"a"}, then the value
  of this attribute will replace \code{"a"} in the \code{show} output.
  If it has a \code{"X"} attribute, then the value of this attribute will
  replace \code{"X"}.
  It is also possible to control the \code{quotientBar} argument by
  assigning a \code{"quotientBar"} attribute to the \code{symbolicQspray}
  object to be printed.
}
\examples{
Qspray <- rSymbolicQspray()
showSymbolicQsprayCanonical(quotientBar = " / ")
}
